Fix missing digital detection & Custom file input

+++ resources/main.cpp
@@ -1,20 +1,15 @@
 #include <stdio.h>
-#include <stdlib.h>
-#include <string.h>
-
 
 int main() {
-
-    long long i, sum = 0;
-    while (scanf("%lld", &i) != EOF) {
+    long i, sum = 0;
+    while (scanf("%ld", &i) != EOF) {
         while (i--) {
-            long long k;
-            scanf("%lld", &k);
+            long k;
+            scanf("%ld", &k);
             sum += k;
         }
-        printf("%lld", sum);
+        printf("%ld", sum);
         sum = 0;
     }
     return 0;
 }
-
